pyproject.toml
[tool.poetry]
name = "toudoum-framework"
version = "0.1.0"
description = ""
authors = ["msterhuj <gabin.lanore@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
Markdown = "^3.4.1"
celery = "^5.2.7"
django-redis = "^5.2.0"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
iptools = "^0.7.0"
gunicorn = "^20.1.0"
requests = "^2.28.1"
ntplib = "^0.4.0"
shodan = "^1.28.0"
scapy = "^2.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
